[Scheduler] Multistep DPM-solver for Python side (#2)
This PR introduces the multistep DPM-solver tracing and the Python side
runtime. With this PR, we can now deploy the stable diffusion locally
with the multistep DPM-solver, which is set to generate an image with a
fixed number (20) of steps. By reducing the number of UNet iterations
from 50 to 20, the image generation is accelerated by 2.5x.

This PR also organizes the code around schedulers, so that the codebase
is more manageable and extensible when more possible schedulers join in
the future.

The web deployment scripts get broken by this PR. They will be fixed
soon when the multistep DPM-solver for web side is introduced.