[Database] Update database with no spatial block loop unroll (#7)
Prior to this PR, the database contains traces with which the PrimFuncs
are scheduled as two kernels, with second one being a loop-unrolled
spatial block. However, spatial blocks are not supposed to be unrolled,
and the issue above is what MetaSchedule doesn’t consider at this
moment.

This PR mutates the database manually to remove those loop unroll for
spatial blocks. The MetaSchedule schedule rule and/or post-processor
should be enhanced in the future.